Changed RGBColor to doubles

+++ raytracer/utilities/RGBColor.hpp
@@ -13,13 +13,13 @@
 
 class RGBColor {
  public:
-  float r, g, b;  // red, green, and blue components in [0,1].
+  double r, g, b;  // red, green, and blue components in [0,1].
 
  public:
   // Constructors.
   RGBColor();                              // set color to (0, 0, 0).
-  explicit RGBColor(float c);              // set color to (c, c, c).
-  RGBColor(float _r, float _g, float _b);  // set color to (_r, _g, _b).
+  explicit RGBColor(double c);              // set color to (c, c, c).
+  RGBColor(double _r, double _g, double _b);  // set color to (_r, _g, _b).
 
   // Copy constructor and assignment operator.
   RGBColor(const RGBColor& c) = default;
@@ -31,18 +31,18 @@ class RGBColor {
   // Operations.
   RGBColor operator+(const RGBColor& c) const;  // addition.
   RGBColor& operator+=(const RGBColor& c);      // compound addition.
-  RGBColor operator*(const float a) const;      // multiplication by a float.
+  RGBColor operator*(const double a) const;      // multiplication by a double.
   RGBColor operator*(
       const RGBColor& c) const;         // component-wise multiplication.
-  RGBColor& operator*=(const float a);  // compound multiplication by a float.
-  RGBColor operator/(const float a) const;   // division by a float.
-  RGBColor& operator/=(const float a);       // compound division by a float.
+  RGBColor& operator*=(const double a);  // compound multiplication by a double.
+  RGBColor operator/(const double a) const;   // division by a double.
+  RGBColor& operator/=(const double a);       // compound division by a double.
   bool operator==(const RGBColor& c) const;  // equality.
-  RGBColor powc(float p) const;              // raise components to a power.
-  float average() const;                     // the average of the components.
+  RGBColor powc(double p) const;              // raise components to a power.
+  double average() const;                     // the average of the components.
 };
 
-// Front multiplication by a float.
-RGBColor operator*(const float a, const RGBColor& c);
+// Front multiplication by a double.
+RGBColor operator*(const double a, const RGBColor& c);
 
 #endif  // RAYTRACER_UTILITIES_RGBCOLOR_HPP_
